Add trace relation to mergeable relation
As a first experiment, add a simple instance of the trace relation to
the mergeable states relation, and repair proofs dealing with the
mergeability of the parts of the state and its symmetry.

The fancy induction principle on mergeable states,, as well as the
proofs that depend on it, remain to be fixed, including the more
questionable (admitted) result on the mergeability of stacks, which
needs more careful revision.