Add third state to mergeability relation
One of the most delicate aspects of the recombination proofs comes from
the necessity to combine a pair of states, corresponding to two parallel
executions, into a new one that corresponds to the derived execution
that combines parts of both original runs. This operation is relatively
straightforward when no sharing is allowed, and becomes predictably
trickier once a program may modify parts of the state that do not belong
to it, as sharing extends joint ownership to various parts of a program.

To attempt to avoid the need for functions that combine states based on
sharing and active components, as well as the need to prove results
about these computations, we extend the state mergeability relation to
make the third, recombined run of the programs explicit, together with
information about the recombined run so far and the trace relation that
is also safisfied with this derived execution.

We comment out proofs about the computational definitions that we hope
to make redundant to focus on the more immediately interesting
simulation results on the modified mergeability relation.