Remove shared addresses from I state and semantics
The set of shared addresses is no longer necessary as part of the state,
so we remove it from it and the semantics. Previously stubbed
definitions related to its computation are no longer needed and also
removed. Some related comments are also removed, while a few notes of
potential interest are added.

This is a partial change, in that the proofs that depend on those
changes still need to be adjusted.