Non-informative in terms of informative semantics
The extant definition of the non-informative semantics is a wrapper of
sorts around the "full" informative semantics, but is rather
repetitious, as each possible instruction in the non-informative
relation wraps an instance of the informative relation, and performs
some implicit simplifications on which events need to be filtered and
which are filtered implicitly.

To simplify both definition and proofs, we can consider replacing that
definition and stating simply that a non-informative step is a wrapper
around an informative step, where the produced trace is filtered.