Some definability fixes, esp. to the _gen lemma
The interesting cases of the proof of [definability_gen] can be
unblocked by a combination of the now-computable definitions and the
observation (to be made precise in-proof) that the result of
dereferencing a register location must result in a value.

The remaining subgoals in the main theorem continue to follow the
existing structure, with a number of subgoals corresponding to the
hypotheses established for those, previously non-computable
definitions.