Stub back-translation initialization code
Not added to back-translation yet (check hardwired to false) while
working on later proofs.

The basic idea is this: the initialization code for a given
component is executed the first time that component is run. This would
be inserted as a prelude to the code of the first procedure called in
said component.