Well-formed source program buffers are pointerless
Refine the definition of well-formed programs to forbid pointer values
not only in well-formed expressions (no pointer forging), but also in
static buffers. As those are not part of the program interface, their
sharing model is unclear, more so given that in the current model only
one static buffer per component can be declared. This sort of sharing
would furthermore complicate the definition of the trace and memory
relations and is best avoided at the moment.

Some uninteresting admits left in a couple of proofs to explore the
effects of the change on back-translation up front.