Informative-only events in [definability_gen]
By adding the recursive calls to the informative-only events, we can
execute their generated expressions as well as get to a point (i.e.,
the recursive call) where it is possible to reestablish the state
well-formedness invariant.

The same strategy seems to work well with a few tweaks for all
outstanding event types.