No pointer values in intermediate static buffers
Strengthen the well-formedness condition on intermediate programs to
ensure that static buffers do not contain any pointer values, i.e.,
initial memory sharing capabilities. This change does not limit the
expressivity of the programs that can be written in the language,
while avoiding significant complications in the trace and memory
relations should those static pointers continue to be allowed.