Merge memory and state/trace relations
The existing proof structure looks like it will adapt fairly well to
the richer setting of memory sharing, but some proofs require more
information than is currently available from the simple memory relation
devised to keep track of that region of the states. By integrating this
memory relation in the standard state mergeability relation and
establishing the necessary connections, we can begin to repair some of
the results that will be necessary to restore the basic building blocks
of the recombination proof.