Lockstep events
Modify theorem statements on lockstep events to specify different but
related instead of identical events.

Having done this, we can finish fixing the general outline of
[threeway_multisem_star_program] without significant changes.

So far memory relations are essentially used as black boxes, assuming
they can be composed and decomposed in predictable ways. At least at
this level there appear to be no problems so far.