Add register relation
Similar to the memory relation, the register relation has the
recombined program track the first program (from which the "program
side" is taken) when said part is executing, and the second program
(from which the "context side" is taken) otherwise.

Some new proof obligations arise from this strengthening of the
simulation invariant.