A couple of memory lemmas
Based on the result of a previous memory operation, predict the success
of a subsequent operation on the resulting memory. Some results are
already available, sometimes restricted to component memories, and
others are yet to be added. The proof of these two lemmas should be
simple, and their statements could be modified to better suit the task
at hand.