Push store helper inside EConst sub-cases
In [definability_gen], the [EConst] event consists of three cases, one
for each class of constant value. However, the value being stored in the
metadata buffer that simulates the destination register is not always
the exact value of the constant; in the case of pointer values, the
corresponding address translation is the value actually being stored.
Consequently, the concrete instance of the well-formed memory lemma must
be adapted to each sub-case.