First version of well-formed memory events
The well-formed memory lemma does not have enough information about the
events it appends at the end of the trace to be capable of restoring the
strengthened invariant, in particular the new components describing the
desired properties of call, return and alloc events. Added events should
follow certain rules, and those rules should be easily provable for the
traces of interest.