Some register invariants of well-formed states
These goals are fairly mechanical, like their memory counterparts. For
the most part, they involve reasoning on the register operations and
considering the various possible cases, notable regarding equality of
registers in sequences on operations, and dealing with those cases in
which new values (i.e., not already in the register file) are written.