Simulate local buffers during back-translation
Move from the outdated offset-based scheme to block-based shifting.
This requires an extension to the metadata buffer to allocate the
buffer that will simulate the original local buffer, as well as
additional initialization code to fill out the newly allocated buffer
with the contents of the original static program buffers. The original
experimental code is modified to do so.

Co-authored-by: Akram El-Korashy <elkorashy@mpi-sws.org>
Co-authored-by: Jérémy Thibault <jeremy.thibault@csp.mpg.de>