Another fix to the back-translation function
The expression that back-translates a load must dereference the source
register twice: once to fetch the memory cell that simulates said
register, and then a second time to fetch from memory the value to load.
The simulated register must moreover contain a pointer, and the second
load is no longer constrained to operate in the same compartment.