Complete statement of initialization lemma
Next block to be allocated does not change for uninitialized components.
The existing invariant almost implies that directly. It should be
possible to prove this outside the lemma making use of the fact that the
semantics does not allow the allocation of empty blocks, and therefore
identical loads imply identical memory blocks.