Extend component memory with alloc invariant
As part of a component memory, maintain the property that the memory
does not contain blocks below the next block identifier to be
allocated.

Proof irrelevance is used to establish equality of component memories.
A small number of proof obligations arise from the change in the
definition. Additionally, case analysis in some proofs needs to be
adjusted owing to the presence of dependent pattern matching.

In passing, comment some experimental but ultimately unused memory
operations and the results on those (these simplifications are not
exhaustive).

Co-authored-by: Adrien Durier <adrien.durier@mpi-sp.org>
Co-authored-by: Akram El-Korashy <elkorashy@mpi-sws.org>
Co-authored-by: Jérémy Thibault <jeremy.thibault@mpi-sp.org>