Axiomatize component memory equality
Use this fact to prove related corollary for definability.

Note that under the current memory model the desired corollary does not
hold in general, although it does for the actual memories used by our
languages. A slight strengthening of the model is required in order for
this to work at this level.